requirement.txt
appdirs==1.4.4
asn1crypto==0.24.0
attrs==20.1.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
bleach==3.1.5
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
entrypoints==0.3
enum34==1.1.6
filelock==3.0.12
functools32==3.2.3.post2
future==0.18.2
futures==3.3.0
idna==2.6
importlib-metadata==1.7.0
importlib-resources==3.0.0
ipaddress==1.0.17
ipykernel==4.10.1
ipython==5.10.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.5
jupyter-console==5.2.0
jupyter-core==4.6.3
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==2.2.5
meld3==1.0.2
mistune==0.8.4
nbconvert==5.6.1
nbformat==4.4.0
notebook==5.7.10
numpy==1.16.6
packaging==20.4
pandocfilters==1.4.2
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.2
prometheus-client==0.8.0
prompt-toolkit==1.0.18
protobuf==3.12.4
ptyprocess==0.6.0
pycrypto==2.6.1
Pygments==2.5.2
pygobject==3.26.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pyxdg==0.25
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
scandir==1.10.0
scipy==1.2.3
SecretStorage==2.3.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.5.4
supervisor==3.3.1
tensorboardX==2.1
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
torchvision==0.5.0
tornado==5.1.1
traitlets==4.3.3
typing==3.7.4.3
virtualenv==20.0.28
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==1.2.0
